auth_service/requirements.txt
pydantic[email]==2.12.4
pydantic-settings==2.12.0
fastapi==0.128.0
jinja2==3.1.6
httpx==0.28.1
python-multipart==0.0.21
sqlalchemy[asyncio]==2.0.44
asyncpg==0.30.0
alembic
psycopg2-binary==2.9.11
bcrypt==5.0.0
python-jose==3.5.0
pyjwt==2.10.1
redis==7.1.0
uvicorn==0.38.0
setuptools==80.9.0
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-sqlalchemy==0.60b1